fix command line issue and sync peer issue

+++ src/models/node.rs
@@ -6,17 +6,17 @@ pub struct Node {
     pub peers: Mutex<Vec<NodeInfo>>,
     pub blockchain: Mutex<Blockchain>,
     pub ip: String,
-    pub port: i32,
+    pub port: u32,
     pub id: Uuid,
 }
 
 impl Node {
-    pub fn new(port: u16) -> Node {
+    pub fn new(port: u32) -> Node {
         Node {
             peers: Mutex::new(Vec::new()),
             blockchain: Mutex::new(Blockchain::new()),
             ip: "[::1]".to_string(),
-            port: port as i32,
+            port: port,
             id: Uuid::new_v4(),
         }
     }
